feat(JitterContext, P5Sketch, Slider, algorithmConfig): enhance parameter management and introduce new algorithms
This commit introduces several enhancements across multiple components. The JitterContext now includes additional visual variation parameters and utilizes useCallback for the updateParams function. The P5Sketch component has been updated to dynamically use localParams for various settings, improving flexibility. The Slider component has been refactored to implement a debounced value update for smoother UI interactions. Additionally, new algorithms for generating Perlin noise fields and fractal trees have been added, along with updates to the algorithm configuration to support new parameters and improve user experience.
