+++ lib/algorithmConfig.ts
@@ -25,10 +25,10 @@ export const algorithmParameters: Record<string, ParameterConfig[]> = {
       key: 'shapeVariety',
       label: 'Shape Variety',
       type: 'slider',
-      min: 0,
+      min: 1,
       max: 8,
       step: 1,
-      defaultValue: 2,
+      defaultValue: 1,
     },
     {
       key: 'sizeVariation',
@@ -71,13 +71,13 @@ export const algorithmParameters: Record<string, ParameterConfig[]> = {
       defaultValue: 4,
     },
     {
-      key: 'shapeVariety',
-      label: 'Shape Variety',
+      key: 'gutter',
+      label: 'Spacing',
       type: 'slider',
       min: 0,
-      max: 6,
+      max: 20,
       step: 1,
-      defaultValue: 2,
+      defaultValue: 5,
     },
     {
       key: 'displacementIntensity',
@@ -124,7 +124,7 @@ export const algorithmParameters: Record<string, ParameterConfig[]> = {
       key: 'shapeVariety',
       label: 'Shape Variety',
       type: 'slider',
-      min: 0,
+      min: 1,
       max: 5,
       step: 1,
       defaultValue: 1,
@@ -164,10 +164,10 @@ export const algorithmParameters: Record<string, ParameterConfig[]> = {
       key: 'shapeVariety',
       label: 'Shape Variety',
       type: 'slider',
-      min: 0,
+      min: 1,
       max: 10,
       step: 1,
-      defaultValue: 3,
+      defaultValue: 1,
     },
     {
       key: 'heightVariation',
@@ -176,7 +176,7 @@ export const algorithmParameters: Record<string, ParameterConfig[]> = {
       min: 0,
       max: 10,
       step: 1,
-      defaultValue: 2,
+      defaultValue: 1,
     },
   ],
 
@@ -339,16 +339,29 @@ export const algorithmParameters: Record<string, ParameterConfig[]> = {
       formatValue: (value) => `${value}°`,
     },
     {
-      key: 'axiom',
-      label: 'Axiom',
-      type: 'text',
-      defaultValue: 'F',
-    },
-    {
-      key: 'rules',
-      label: 'Rules',
-      type: 'text',
-      defaultValue: 'F=F+F-F-F+F',
+      key: 'pattern',
+      label: 'L-System Pattern',
+      type: 'select',
+      options: [
+        { value: 'koch', label: 'Koch Curve (F=F+F-F-F+F)' },
+        { value: 'dragon', label: 'Dragon Curve (X=X+YF+, Y=-FX-Y)' },
+        {
+          value: 'sierpinski',
+          label: 'Sierpinski Triangle (F=F-G+F+G-F, G=GG)',
+        },
+        { value: 'plant', label: 'Plant Growth (X=F+[[X]-X]-F[-FX]+X, F=FF)' },
+        { value: 'tree', label: 'Binary Tree (F=F[+F]F[-F]F)' },
+        { value: 'levy', label: 'Levy C Curve (F=+F--F+)' },
+        {
+          value: 'hilbert',
+          label: 'Hilbert Curve (A=-BF+AFA+FB-, B=+AF-BFB-FA+)',
+        },
+        {
+          value: 'gosper',
+          label: 'Gosper Island (A=A-B--B+A++AA+B-, B=+A-BB--B-A++A+B)',
+        },
+      ],
+      defaultValue: 'koch',
     },
   ],
 }
