+++ lib/engine/basicAlgorithms.ts
@@ -1,0 +1,895 @@
+import type { Layer, DrawingCommand, Point } from '@/lib/types'
+import {
+  getCanvasSize,
+  getRandomColor,
+  createColorPalette,
+  perlinNoise,
+} from './algorithmUtils'
+
+export const generateUniformGrid = (layer: Layer): DrawingCommand[] => {
+  const commands: DrawingCommand[] = []
+  const {
+    density,
+    gutter = 5,
+    colorPalette,
+    shapeVariety = 1,
+    sizeVariation = 3,
+  } = layer.parameters
+  const { width: CANVAS_WIDTH, height: CANVAS_HEIGHT } = getCanvasSize(layer)
+
+  const baseCellSize = Math.max(10, 100 - density * 0.8)
+  const spacing = baseCellSize + gutter
+
+  const cols = Math.floor(CANVAS_WIDTH / spacing)
+  const rows = Math.floor(CANVAS_HEIGHT / spacing)
+
+  const offsetX = (CANVAS_WIDTH - cols * spacing) / 2
+  const offsetY = (CANVAS_HEIGHT - rows * spacing) / 2
+
+  // Enhanced color palette
+  const colors = createColorPalette(colorPalette)
+
+  for (let row = 0; row < rows; row++) {
+    for (let col = 0; col < cols; col++) {
+      const baseX = offsetX + col * spacing + baseCellSize / 2
+      const baseY = offsetY + row * spacing + baseCellSize / 2
+
+      // Apply size variation
+      const sizeMultiplier = 1 + (Math.random() - 0.5) * (sizeVariation / 10)
+      const cellSize = baseCellSize * Math.max(0.3, sizeMultiplier)
+
+      // Determine shape based on variety parameter
+      const shapeChoice = Math.floor(Math.random() * shapeVariety)
+
+      // Color selection
+      const primaryColor = getRandomColor(colors)
+      const secondaryColor = getRandomColor(colors)
+
+      switch (shapeChoice) {
+        case 0: // Rectangle (original)
+          commands.push({
+            type: 'rect',
+            position: { x: baseX, y: baseY },
+            size: { width: cellSize, height: cellSize },
+            color: primaryColor,
+            strokeColor: '#333333',
+            strokeWeight: Math.max(0.5, cellSize / 30),
+            alpha: 0.8 + Math.random() * 0.2,
+          })
+          break
+
+        case 1: // Circle
+          commands.push({
+            type: 'ellipse',
+            position: { x: baseX, y: baseY },
+            radius: cellSize / 2,
+            color: primaryColor,
+            strokeColor: '#333333',
+            strokeWeight: Math.max(0.5, cellSize / 30),
+            alpha: 0.8 + Math.random() * 0.2,
+          })
+          break
+
+        case 2: // Triangle
+          const trianglePoints: Point[] = [
+            { x: baseX, y: baseY - cellSize / 2 },
+            { x: baseX - cellSize / 2, y: baseY + cellSize / 2 },
+            { x: baseX + cellSize / 2, y: baseY + cellSize / 2 },
+          ]
+
+          commands.push({
+            type: 'polygon',
+            position: { x: baseX, y: baseY },
+            points: trianglePoints,
+            color: primaryColor,
+            strokeColor: '#333333',
+            strokeWeight: Math.max(0.5, cellSize / 30),
+            alpha: 0.8 + Math.random() * 0.2,
+          })
+          break
+
+        case 3: // Diamond
+          const diamondPoints: Point[] = [
+            { x: baseX, y: baseY - cellSize / 2 },
+            { x: baseX + cellSize / 2, y: baseY },
+            { x: baseX, y: baseY + cellSize / 2 },
+            { x: baseX - cellSize / 2, y: baseY },
+          ]
+
+          commands.push({
+            type: 'polygon',
+            position: { x: baseX, y: baseY },
+            points: diamondPoints,
+            color: primaryColor,
+            strokeColor: '#333333',
+            strokeWeight: Math.max(0.5, cellSize / 30),
+            alpha: 0.8 + Math.random() * 0.2,
+          })
+          break
+
+        case 4: // Hexagon
+          const hexagonPoints: Point[] = []
+          for (let i = 0; i < 6; i++) {
+            const angle = (i * Math.PI * 2) / 6
+            hexagonPoints.push({
+              x: baseX + (Math.cos(angle) * cellSize) / 2,
+              y: baseY + (Math.sin(angle) * cellSize) / 2,
+            })
+          }
+
+          commands.push({
+            type: 'polygon',
+            position: { x: baseX, y: baseY },
+            points: hexagonPoints,
+            color: primaryColor,
+            strokeColor: '#333333',
+            strokeWeight: Math.max(0.5, cellSize / 30),
+            alpha: 0.8 + Math.random() * 0.2,
+          })
+          break
+
+        case 5: // Plus/Cross
+          // Vertical bar
+          commands.push({
+            type: 'rect',
+            position: { x: baseX, y: baseY },
+            size: { width: cellSize / 3, height: cellSize },
+            color: primaryColor,
+            strokeColor: '#333333',
+            strokeWeight: Math.max(0.5, cellSize / 40),
+            alpha: 0.8 + Math.random() * 0.2,
+          })
+
+          // Horizontal bar
+          commands.push({
+            type: 'rect',
+            position: { x: baseX, y: baseY },
+            size: { width: cellSize, height: cellSize / 3 },
+            color: secondaryColor,
+            strokeColor: '#333333',
+            strokeWeight: Math.max(0.5, cellSize / 40),
+            alpha: 0.8 + Math.random() * 0.2,
+          })
+          break
+
+        case 6: // Star
+          const starPoints: Point[] = []
+          for (let i = 0; i < 10; i++) {
+            const angle = (i * Math.PI * 2) / 10
+            const radius = i % 2 === 0 ? cellSize / 2 : cellSize / 4
+            starPoints.push({
+              x: baseX + Math.cos(angle) * radius,
+              y: baseY + Math.sin(angle) * radius,
+            })
+          }
+
+          commands.push({
+            type: 'polygon',
+            position: { x: baseX, y: baseY },
+            points: starPoints,
+            color: primaryColor,
+            strokeColor: '#333333',
+            strokeWeight: Math.max(0.5, cellSize / 30),
+            alpha: 0.8 + Math.random() * 0.2,
+          })
+          break
+      }
+    }
+  }
+
+  return commands
+}
+
+export const generateNoiseDisplacementGrid = (
+  layer: Layer
+): DrawingCommand[] => {
+  const commands: DrawingCommand[] = []
+  const {
+    density,
+    gutter = 5,
+    colorPalette,
+    displacementIntensity = 1,
+    noiseScale = 0.005,
+    octaves = 4,
+  } = layer.parameters
+  const { width: CANVAS_WIDTH, height: CANVAS_HEIGHT } = getCanvasSize(layer)
+
+  // More responsive density scaling
+  const baseCellSize = Math.max(
+    5,
+    Math.min(CANVAS_WIDTH, CANVAS_HEIGHT) / (5 + density * 1.5)
+  )
+  const spacing = baseCellSize + gutter
+
+  // More responsive displacement scaling
+  const displacementAmount = Math.max(
+    5,
+    baseCellSize * 0.8 * displacementIntensity
+  )
+
+  const cols = Math.floor(CANVAS_WIDTH / spacing)
+  const rows = Math.floor(CANVAS_HEIGHT / spacing)
+
+  const offsetX = (CANVAS_WIDTH - cols * spacing) / 2
+  const offsetY = (CANVAS_HEIGHT - rows * spacing) / 2
+
+  // Enhanced color palette
+  const colors = createColorPalette(colorPalette)
+
+  // More responsive noise scale - use actual parameter value more directly
+  const actualNoiseScale = Math.max(0.001, noiseScale)
+
+  // Multi-octave noise function with better parameter responsiveness
+  const multiOctaveNoise = (x: number, y: number): number => {
+    let value = 0
+    let amplitude = 1
+    let frequency = actualNoiseScale
+
+    for (let i = 0; i < octaves; i++) {
+      value += perlinNoise.noise2D(x * frequency, y * frequency) * amplitude
+      amplitude *= 0.5
+      frequency *= 2
+    }
+
+    return value
+  }
+
+  for (let row = 0; row < rows; row++) {
+    for (let col = 0; col < cols; col++) {
+      const baseX = offsetX + col * spacing + baseCellSize / 2
+      const baseY = offsetY + row * spacing + baseCellSize / 2
+
+      // Apply noise-based displacement with better parameter responsiveness
+      const noiseInputX = baseX * actualNoiseScale
+      const noiseInputY = baseY * actualNoiseScale
+
+      const noiseX =
+        multiOctaveNoise(noiseInputX, noiseInputY) * displacementAmount
+      const noiseY =
+        multiOctaveNoise(noiseInputX + 1000, noiseInputY + 1000) *
+        displacementAmount
+
+      const x = baseX + noiseX
+      const y = baseY + noiseY
+
+      // Vary size based on noise with more variation
+      const sizeNoiseInput = actualNoiseScale * 0.5
+      const sizeNoise = multiOctaveNoise(
+        baseX * sizeNoiseInput,
+        baseY * sizeNoiseInput
+      )
+      const size = baseCellSize * (0.4 + Math.abs(sizeNoise) * 1.2)
+
+      // Color selection based on position and noise
+      const colorNoiseInput = actualNoiseScale * 0.3
+      const colorNoise = multiOctaveNoise(
+        baseX * colorNoiseInput,
+        baseY * colorNoiseInput
+      )
+      const colorIndex =
+        Math.floor(Math.abs(colorNoise) * colors.length) % colors.length
+      const primaryColor = colors[colorIndex]
+
+      // Create organic blob shapes displaced by noise
+      const blobPoints: Point[] = []
+      const numPoints = Math.max(3, 6 + Math.floor(Math.abs(sizeNoise) * 6))
+
+      for (let i = 0; i < numPoints; i++) {
+        const angle = (i * Math.PI * 2) / numPoints
+
+        // Use more responsive noise input for radius variation
+        const radiusNoiseInput = actualNoiseScale * 2
+        const radiusVariation =
+          0.6 +
+          Math.abs(
+            multiOctaveNoise(
+              (baseX + Math.cos(angle) * 10) * radiusNoiseInput,
+              (baseY + Math.sin(angle) * 10) * radiusNoiseInput
+            )
+          ) *
+            0.8
+
+        const radius = (size / 2) * radiusVariation
+        blobPoints.push({
+          x: x + Math.cos(angle) * radius,
+          y: y + Math.sin(angle) * radius,
+        })
+      }
+
+      commands.push({
+        type: 'polygon',
+        position: { x, y },
+        points: blobPoints,
+        color: primaryColor,
+        strokeColor: '#555555',
+        strokeWeight: Math.max(0.3, size / 25),
+        alpha: 0.6 + Math.abs(sizeNoise) * 0.4,
+      })
+    }
+  }
+
+  return commands
+}
+
+export const generateRecursiveSubdivision = (
+  layer: Layer
+): DrawingCommand[] => {
+  const commands: DrawingCommand[] = []
+  const {
+    density,
+    colorPalette,
+    subdivisions = 3,
+    threshold = 0.5,
+    shapeVariety = 1,
+    colorVariation = 3,
+  } = layer.parameters
+  const { width: CANVAS_WIDTH, height: CANVAS_HEIGHT } = getCanvasSize(layer)
+
+  // Calculate subdivision parameters - allow deeper subdivisions
+  const maxDepth = Math.max(2, Math.min(subdivisions, 12))
+  const minSize = Math.max(
+    8, // Reduced minimum size to allow more subdivisions
+    Math.min(CANVAS_WIDTH, CANVAS_HEIGHT) / (10 + density / 3)
+  )
+
+  // Create regions to track subdivisions
+  interface Region {
+    x: number
+    y: number
+    width: number
+    height: number
+    depth: number
+  }
+
+  const regions: Region[] = []
+
+  // Start with the full canvas
+  regions.push({
+    x: 0,
+    y: 0,
+    width: CANVAS_WIDTH,
+    height: CANVAS_HEIGHT,
+    depth: 0,
+  })
+
+  // Process regions iteratively to avoid recursion issues
+  while (regions.length > 0) {
+    const region = regions.shift()!
+
+    // Check if we should subdivide this region
+    // Force subdivision for early depths, then use probability
+    let shouldSubdivide = false
+
+    if (
+      region.depth < maxDepth &&
+      region.width > minSize &&
+      region.height > minSize
+    ) {
+      if (region.depth < 2) {
+        // Force subdivision for first 2 levels to ensure we get some structure
+        shouldSubdivide = true
+      } else {
+        // Use threshold for deeper levels - higher threshold = more subdivisions
+        const subdivisionChance = threshold * (1 - region.depth * 0.1)
+        shouldSubdivide = Math.random() < Math.max(0.2, subdivisionChance)
+      }
+    }
+
+    if (!shouldSubdivide) {
+      // Draw this region with shape variety
+      const padding = Math.max(2, region.width * 0.05, region.height * 0.05)
+      const centerX = region.x + region.width / 2
+      const centerY = region.y + region.height / 2
+      const shapeWidth = region.width - padding
+      const shapeHeight = region.height - padding
+
+      // Enhanced color palette with variation
+      const colors = createColorPalette(colorPalette)
+      const colorIndex =
+        (region.x + region.y + region.depth * colorVariation) % colors.length
+      const primaryColor = colors[Math.floor(colorIndex)]
+      const secondaryColor =
+        colors[Math.floor((colorIndex + 1) % colors.length)]
+
+      // Shape choice based on region properties and variety setting
+      const shapeChoice =
+        Math.floor((region.x + region.y + region.depth) * 0.01) % shapeVariety
+
+      switch (shapeChoice) {
+        case 0: // Rectangle (original)
+          commands.push({
+            type: 'rect',
+            position: { x: centerX, y: centerY },
+            size: { width: shapeWidth, height: shapeHeight },
+            color: primaryColor,
+            strokeColor: '#000000',
+            strokeWeight: Math.max(0.5, 2 - region.depth * 0.3),
+            alpha: 0.7 + Math.random() * 0.3,
+          })
+          break
+
+        case 1: // Ellipse
+          const radius = Math.min(shapeWidth, shapeHeight) / 2
+          commands.push({
+            type: 'ellipse',
+            position: { x: centerX, y: centerY },
+            radius: radius,
+            color: primaryColor,
+            strokeColor: '#000000',
+            strokeWeight: Math.max(0.5, 2 - region.depth * 0.3),
+            alpha: 0.7 + Math.random() * 0.3,
+          })
+          break
+
+        case 2: // Triangle
+          const trianglePoints: Point[] = [
+            { x: centerX, y: centerY - shapeHeight / 2 },
+            { x: centerX - shapeWidth / 2, y: centerY + shapeHeight / 2 },
+            { x: centerX + shapeWidth / 2, y: centerY + shapeHeight / 2 },
+          ]
+
+          commands.push({
+            type: 'polygon',
+            position: { x: centerX, y: centerY },
+            points: trianglePoints,
+            color: primaryColor,
+            strokeColor: '#000000',
+            strokeWeight: Math.max(0.5, 2 - region.depth * 0.3),
+            alpha: 0.7 + Math.random() * 0.3,
+          })
+          break
+
+        case 3: // Diamond
+          const diamondPoints: Point[] = [
+            { x: centerX, y: centerY - shapeHeight / 2 },
+            { x: centerX + shapeWidth / 2, y: centerY },
+            { x: centerX, y: centerY + shapeHeight / 2 },
+            { x: centerX - shapeWidth / 2, y: centerY },
+          ]
+
+          commands.push({
+            type: 'polygon',
+            position: { x: centerX, y: centerY },
+            points: diamondPoints,
+            color: primaryColor,
+            strokeColor: '#000000',
+            strokeWeight: Math.max(0.5, 2 - region.depth * 0.3),
+            alpha: 0.7 + Math.random() * 0.3,
+          })
+          break
+
+        case 4: // Cross/Plus
+          // Vertical bar
+          commands.push({
+            type: 'rect',
+            position: { x: centerX, y: centerY },
+            size: { width: shapeWidth / 3, height: shapeHeight },
+            color: primaryColor,
+            strokeColor: '#000000',
+            strokeWeight: Math.max(0.3, 1.5 - region.depth * 0.2),
+            alpha: 0.8,
+          })
+
+          // Horizontal bar
+          commands.push({
+            type: 'rect',
+            position: { x: centerX, y: centerY },
+            size: { width: shapeWidth, height: shapeHeight / 3 },
+            color: secondaryColor,
+            strokeColor: '#000000',
+            strokeWeight: Math.max(0.3, 1.5 - region.depth * 0.2),
+            alpha: 0.8,
+          })
+          break
+      }
+
+      continue
+    }
+
+    // Decide how to split this region
+    const aspectRatio = region.width / region.height
+    let splitVertically: boolean
+
+    if (aspectRatio > 1.5) {
+      // Wide rectangle - prefer vertical split
+      splitVertically = true
+    } else if (aspectRatio < 0.67) {
+      // Tall rectangle - prefer horizontal split
+      splitVertically = false
+    } else {
+      // Square-ish - random choice
+      splitVertically = Math.random() > 0.5
+    }
+
+    if (splitVertically) {
+      // Split vertically (left/right)
+      const splitRatio = 0.3 + Math.random() * 0.4 // 30% to 70%
+      const splitPosition = region.width * splitRatio
+
+      // Left region
+      regions.push({
+        x: region.x,
+        y: region.y,
+        width: splitPosition,
+        height: region.height,
+        depth: region.depth + 1,
+      })
+
+      // Right region
+      regions.push({
+        x: region.x + splitPosition,
+        y: region.y,
+        width: region.width - splitPosition,
+        height: region.height,
+        depth: region.depth + 1,
+      })
+    } else {
+      // Split horizontally (top/bottom)
+      const splitRatio = 0.3 + Math.random() * 0.4 // 30% to 70%
+      const splitPosition = region.height * splitRatio
+
+      // Top region
+      regions.push({
+        x: region.x,
+        y: region.y,
+        width: region.width,
+        height: splitPosition,
+        depth: region.depth + 1,
+      })
+
+      // Bottom region
+      regions.push({
+        x: region.x,
+        y: region.y + splitPosition,
+        width: region.width,
+        height: region.height - splitPosition,
+        depth: region.depth + 1,
+      })
+    }
+  }
+
+  return commands
+}
+
+export const generateIsometricGrid = (layer: Layer): DrawingCommand[] => {
+  const commands: DrawingCommand[] = []
+  const {
+    density,
+    colorPalette,
+    perspective = 0.5,
+    shapeVariety = 1,
+    heightVariation = 2,
+  } = layer.parameters
+  const { width: CANVAS_WIDTH, height: CANVAS_HEIGHT } = getCanvasSize(layer)
+
+  // Scale cell size based on canvas and density
+  const baseCellSize = Math.max(10, Math.min(CANVAS_WIDTH, CANVAS_HEIGHT) / 15)
+  const cellSize = Math.max(8, baseCellSize * (1 - density / 200))
+
+  // Use perspective parameter to control isometric perspective
+  // Clamp perspective more aggressively to prevent grid breakdown
+  const clampedPerspective = Math.min(Math.max(perspective, 0), 1.2)
+  const perspectiveAngle = Math.PI / 12 + (clampedPerspective * Math.PI) / 6 // 15° to 45°
+  const isoWidth = cellSize * Math.max(0.3, Math.cos(perspectiveAngle))
+  const isoHeight = cellSize * Math.sin(perspectiveAngle)
+
+  // Add perspective scaling with clamped value
+  const perspectiveScale = 1 + clampedPerspective * 0.3
+  const scaledIsoWidth = isoWidth * perspectiveScale
+  const scaledIsoHeight = isoHeight * perspectiveScale
+
+  // Calculate grid spacing with safety checks
+  const gridSpacingX = Math.max(20, scaledIsoWidth * 1.8)
+  const gridSpacingY = Math.max(20, scaledIsoHeight * 1.6)
+  const cols = Math.min(
+    30,
+    Math.max(1, Math.ceil(CANVAS_WIDTH / gridSpacingX) + 1)
+  )
+  const rows = Math.min(
+    30,
+    Math.max(1, Math.ceil(CANVAS_HEIGHT / gridSpacingY) + 1)
+  )
+
+  // Center the grid
+  const offsetX = (CANVAS_WIDTH - (cols - 1) * gridSpacingX) / 2
+  const offsetY = (CANVAS_HEIGHT - (rows - 1) * gridSpacingY) / 2
+
+  // Shape generation functions
+  const createIsometricCube = (
+    centerX: number,
+    centerY: number,
+    width: number,
+    height: number,
+    depth: number
+  ) => {
+    const topFace: Point[] = [
+      { x: centerX, y: centerY - height },
+      { x: centerX + width, y: centerY - height + depth },
+      { x: centerX + width - width, y: centerY - height + depth + width },
+      { x: centerX - width, y: centerY - height + width },
+    ]
+
+    const leftFace: Point[] = [
+      { x: centerX - width, y: centerY - height + width },
+      { x: centerX - width, y: centerY + width },
+      { x: centerX, y: centerY + height },
+      { x: centerX, y: centerY - height },
+    ]
+
+    const rightFace: Point[] = [
+      { x: centerX, y: centerY - height },
+      { x: centerX, y: centerY + height },
+      { x: centerX + width, y: centerY + height + depth },
+      { x: centerX + width, y: centerY - height + depth },
+    ]
+
+    return { topFace, leftFace, rightFace }
+  }
+
+  const createIsometricPyramid = (
+    centerX: number,
+    centerY: number,
+    width: number,
+    height: number
+  ) => {
+    const base: Point[] = [
+      { x: centerX - width, y: centerY },
+      { x: centerX, y: centerY - width * 0.5 },
+      { x: centerX + width, y: centerY },
+      { x: centerX, y: centerY + width * 0.5 },
+    ]
+
+    const apex = { x: centerX, y: centerY - height }
+
+    return { base, apex }
+  }
+
+  const createIsometricCylinder = (
+    centerX: number,
+    centerY: number,
+    radius: number,
+    height: number
+  ) => {
+    const ellipseTop = {
+      x: centerX,
+      y: centerY - height,
+      radiusX: radius,
+      radiusY: radius * 0.3,
+    }
+    const ellipseBottom = {
+      x: centerX,
+      y: centerY,
+      radiusX: radius,
+      radiusY: radius * 0.3,
+    }
+
+    const sides: Point[] = [
+      { x: centerX - radius, y: centerY - height + radius * 0.3 },
+      { x: centerX - radius, y: centerY - radius * 0.3 },
+      { x: centerX + radius, y: centerY - radius * 0.3 },
+      { x: centerX + radius, y: centerY - height + radius * 0.3 },
+    ]
+
+    return { ellipseTop, ellipseBottom, sides }
+  }
+
+  for (let row = 0; row < rows; row++) {
+    for (let col = 0; col < cols; col++) {
+      const x = offsetX + col * gridSpacingX + (row % 2) * gridSpacingX * 0.5
+      const y = offsetY + row * gridSpacingY
+
+      // Skip if outside canvas bounds
+      if (
+        x < -scaledIsoWidth ||
+        x > CANVAS_WIDTH + scaledIsoWidth ||
+        y < -scaledIsoHeight ||
+        y > CANVAS_HEIGHT + scaledIsoHeight
+      )
+        continue
+
+      const centerX = x
+      const centerY = y
+
+      // Determine shape based on shapeVariety parameter
+      const shapeChoice = Math.floor(Math.random() * shapeVariety)
+
+      // Apply height variation
+      const heightMultiplier = 1 + (Math.random() - 0.5) * (heightVariation / 5)
+      const currentHeight = scaledIsoHeight * heightMultiplier
+      const currentWidth = scaledIsoWidth * (0.8 + Math.random() * 0.4)
+
+      // Color variations
+      const colors = createColorPalette(colorPalette)
+      const primaryColor = getRandomColor(colors)
+      const secondaryColor = getRandomColor(colors)
+      const shadowColor = '#333333'
+
+      switch (shapeChoice) {
+        case 0: // Hexagon (original)
+          const perspectiveDistortion = 1 + (clampedPerspective - 0.5) * 0.6
+          const hexPoints: Point[] = [
+            { x: centerX, y: centerY - currentHeight * perspectiveDistortion },
+            {
+              x: centerX + currentWidth,
+              y: centerY - currentHeight * 0.5 * perspectiveDistortion,
+            },
+            {
+              x: centerX + currentWidth,
+              y: centerY + currentHeight * 0.5 * perspectiveDistortion,
+            },
+            { x: centerX, y: centerY + currentHeight * perspectiveDistortion },
+            {
+              x: centerX - currentWidth,
+              y: centerY + currentHeight * 0.5 * perspectiveDistortion,
+            },
+            {
+              x: centerX - currentWidth,
+              y: centerY - currentHeight * 0.5 * perspectiveDistortion,
+            },
+          ]
+
+          commands.push({
+            type: 'polygon',
+            position: { x: centerX, y: centerY },
+            points: hexPoints,
+            color: primaryColor,
+            strokeColor: shadowColor,
+            strokeWeight: 1,
+            alpha: 0.8 + Math.random() * 0.2,
+          })
+          break
+
+        case 1: // Isometric cube
+          const cube = createIsometricCube(
+            centerX,
+            centerY,
+            currentWidth * 0.6,
+            currentHeight,
+            currentWidth * 0.3
+          )
+
+          // Draw cube faces with different shades
+          commands.push({
+            type: 'polygon',
+            position: { x: centerX, y: centerY },
+            points: cube.topFace,
+            color: primaryColor,
+            strokeColor: shadowColor,
+            strokeWeight: 1,
+            alpha: 0.9,
+          })
+
+          commands.push({
+            type: 'polygon',
+            position: { x: centerX, y: centerY },
+            points: cube.leftFace,
+            color: secondaryColor,
+            strokeColor: shadowColor,
+            strokeWeight: 1,
+            alpha: 0.7,
+          })
+
+          commands.push({
+            type: 'polygon',
+            position: { x: centerX, y: centerY },
+            points: cube.rightFace,
+            color: primaryColor,
+            strokeColor: shadowColor,
+            strokeWeight: 1,
+            alpha: 0.6,
+          })
+          break
+
+        case 2: // Triangle/Pyramid
+          const pyramid = createIsometricPyramid(
+            centerX,
+            centerY,
+            currentWidth,
+            currentHeight
+          )
+
+          commands.push({
+            type: 'polygon',
+            position: { x: centerX, y: centerY },
+            points: pyramid.base,
+            color: secondaryColor,
+            strokeColor: shadowColor,
+            strokeWeight: 1,
+            alpha: 0.7,
+          })
+
+          // Draw pyramid sides as triangles
+          for (let i = 0; i < pyramid.base.length; i++) {
+            const nextI = (i + 1) % pyramid.base.length
+            const trianglePoints = [
+              pyramid.base[i],
+              pyramid.base[nextI],
+              pyramid.apex,
+            ]
+
+            commands.push({
+              type: 'polygon',
+              position: { x: centerX, y: centerY },
+              points: trianglePoints,
+              color: i % 2 === 0 ? primaryColor : secondaryColor,
+              strokeColor: shadowColor,
+              strokeWeight: 1,
+              alpha: 0.8,
+            })
+          }
+          break
+
+        case 3: // Circle/Ellipse
+          commands.push({
+            type: 'ellipse',
+            position: { x: centerX, y: centerY },
+            radius: currentWidth,
+            color: primaryColor,
+            strokeColor: shadowColor,
+            strokeWeight: 1,
+            alpha: 0.8,
+          })
+          break
+
+        case 4: // Cylinder (only for shapeVariety >= 4)
+          const cylinder = createIsometricCylinder(
+            centerX,
+            centerY,
+            currentWidth * 0.6,
+            currentHeight
+          )
+
+          // Draw cylinder parts
+          commands.push({
+            type: 'polygon',
+            position: { x: centerX, y: centerY },
+            points: cylinder.sides,
+            color: secondaryColor,
+            strokeColor: shadowColor,
+            strokeWeight: 1,
+            alpha: 0.6,
+          })
+
+          commands.push({
+            type: 'ellipse',
+            position: {
+              x: cylinder.ellipseBottom.x,
+              y: cylinder.ellipseBottom.y,
+            },
+            radius: cylinder.ellipseBottom.radiusX,
+            color: primaryColor,
+            strokeColor: shadowColor,
+            strokeWeight: 1,
+            alpha: 0.8,
+          })
+
+          commands.push({
+            type: 'ellipse',
+            position: { x: cylinder.ellipseTop.x, y: cylinder.ellipseTop.y },
+            radius: cylinder.ellipseTop.radiusX,
+            color: primaryColor,
+            strokeColor: shadowColor,
+            strokeWeight: 1,
+            alpha: 0.9,
+          })
+          break
+      }
+    }
+  }
+
+  // Safety check: if no commands generated for isometric grid, add a fallback shape
+  if (commands.length === 0) {
+    const colors = createColorPalette(colorPalette)
+    commands.push({
+      type: 'rect',
+      position: { x: CANVAS_WIDTH / 2, y: CANVAS_HEIGHT / 2 },
+      size: { width: 50, height: 50 },
+      color: getRandomColor(colors),
+      strokeColor: '#333333',
+      strokeWeight: 1,
+      alpha: 0.8,
+    })
+  }
+
+  return commands
+}
